[ENH] Query epochs depending on `level`
